==> notebooks/01_filter_paper.ipynb <==
import json

with open('data/pwc/papers-with-abstracts.json', 'r') as f:
    papers_data = json.load(f)

Only keep the papers that have methods belonging to only one area and this area is not 'General'

filtered_data = []

for paper in papers_data:
    if 'methods' in paper:
        areas = set()

        for method in paper['methods']:
            if 'main_collection' in method and method['main_collection']:
                if 'area' in method['main_collection']:
                    areas.add(method['main_collection']['area'])
        
        if len(areas) == 1 and list(areas)[0] != 'General':
            filtered_data.append(paper)

len(filtered_data)

len(filtered_data)

output_file = 'data/filtered-papers-with-abstracts.json'
with open(output_file, 'w') as outfile:
    json.dump(filtered_data, outfile, indent=4)
